requirements.txt
openai
pyyaml
requests
rich
